(refactor) Update relative paths for navbar
- Also have settings be closer to signup
