+++ client/landingIndex.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Landing from './LandingPage/Landing.jsx';
+import Landing from './LandingComponents/Landing.jsx';
 
 window.onload = () => {
   ReactDOM.render(<Landing />, document.getElementById('landing'));
