+++ client/signupIndex.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import SignUp from './LandingPage/SignUp.jsx';
+import SignUp from './LandingComponents/SignUp.jsx';
 
 window.onload = () => {
   ReactDOM.render(<SignUp />, document.getElementById('signup')
